move tests to subirectories + add support for subdir global check
